TRIPLEX-830 Доработка сторей ChipIcon, ChipMultiselect, ChipOptions, ChipSelect, ChipSort, ChipSuggest
