+++ stories/Chips/ChipMultiselect.stories.tsx
@@ -1,29 +1,32 @@
 import React, { useLayoutEffect, useMemo, useRef, useState } from "react";
-import { MultiselectField } from "../../src/components/MultiselectField";
-import { StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react";
 import {
+    MultiselectField,
     DropdownMobileHeader,
     DropdownMobileInput,
     DropdownMobileClose,
     DropdownMobileBody,
     DropdownMobileFooter,
-} from "../../src/components/Dropdown";
-import { Button, EButtonTheme } from "../../src/components/Button";
-import { EComponentSize } from "../../src/enums/EComponentSize";
-import {
+    Button,
+    EButtonTheme,
+    EComponentSize,
     FormField,
     FormFieldLabel,
     FormFieldInput,
     FormFieldPostfix,
     FormFieldClear,
-} from "../../src/components/FormField";
-import { Checkbox, CheckboxYGroup } from "../../src/components/Checkbox";
-import { ChipMultiselect, IChipMultiselectProps } from "../../src/components/Chip";
+    Checkbox,
+    CheckboxYGroup,
+    ChipMultiselect,
+    ETextSize,
+    Text,
+    EScreenWidth,
+    type ISelectExtendedFieldDropdownProvideProps,
+} from "@sberbusiness/triplex-next";
 import { Title, Description, Primary, Controls, Stories, ArgTypes, Heading } from "@storybook/addon-docs/blocks";
-import { ETextSize, Text } from "../../src/components/Typography";
-import { EScreenWidth } from "../../src/helpers/breakpoints";
 import { AdaptiveUtils } from "../utils/adaptiveUtils";
 import {
+    PlaygroundExample,
     DefaultExample,
     DefaultExampleSource,
     SizesExample,
@@ -31,29 +34,10 @@ import {
     LoadingExample,
     LoadingExampleSource,
 } from "./examples/ChipMultiselect";
+import { CHIP_MULTISELECT_VISUAL_TEST_OPTIONS } from "./examples/ChipMultiselect/storyConstants";
 import "./ChipMultiselect.less";
 
-interface IChipMultiselectOption {
-    id: string;
-    label: string;
-}
-
-const CHIP_MULTISELECT_OPTIONS: IChipMultiselectOption[] = [
-    { id: "multiselect-option-1-1", label: "Значение 1-1" },
-    { id: "multiselect-option-1-2", label: "Значение 1-2" },
-    { id: "multiselect-option-1-3", label: "Значение 1-3" },
-    { id: "multiselect-option-2-1", label: "Значение 2-1" },
-    { id: "multiselect-option-2-2", label: "Значение 2-2" },
-    { id: "multiselect-option-3", label: "Значение 3" },
-];
-
-const CHIP_MULTISELECT_VISUAL_TEST_OPTIONS: IChipMultiselectOption[] = [
-    { id: "multiselect-option-1-1", label: "Значение 1-1" },
-    { id: "multiselect-option-2-1", label: "Значение 2-1" },
-    { id: "multiselect-option-3", label: "Значение 3" },
-];
-
-export default {
+const meta = {
     title: "Components/Chips/ChipMultiselect",
     component: ChipMultiselect,
     tags: ["autodocs"],
@@ -78,9 +62,13 @@ export default {
             ),
         },
     },
-};
+} satisfies Meta<typeof ChipMultiselect>;
 
-export const Playground: StoryObj<typeof ChipMultiselect> = {
+export default meta;
+
+type Story = StoryObj<typeof ChipMultiselect>;
+
+export const Playground: Story = {
     tags: ["!autodocs"],
     args: {
         size: EComponentSize.MD,
@@ -127,199 +115,49 @@ export const Playground: StoryObj<typeof ChipMultiselect> = {
         },
         testRunner: { skip: true },
     },
-    render: (args) => {
-        function createMultiselectFieldStoriesLogic(args: IChipMultiselectProps) {
-            const [selectedIds, setSelectedIds] = useState<string[]>([]);
-            const [filter, setFilter] = useState("");
-            const selectedSet = useMemo(() => new Set(selectedIds), [selectedIds]);
-            const visibleOptions = useMemo(() => {
-                const lower = filter.trim().toLowerCase();
-                if (!lower.length) {
-                    return CHIP_MULTISELECT_OPTIONS;
-                }
-
-                return CHIP_MULTISELECT_OPTIONS.filter((opt) => opt.label.toLowerCase().includes(lower));
-            }, [filter]);
-
-            const handleToggle = (optionId: string, checked: boolean) => {
-                setSelectedIds((prev) => {
-                    if (checked) {
-                        if (prev.includes(optionId)) {
-                            return prev;
-                        }
-
-                        return [...prev, optionId];
-                    }
-
-                    return prev.filter((id) => id !== optionId);
-                });
-            };
-
-            const renderCheckboxList = () => (
-                <CheckboxYGroup aria-label="Options">
-                    {visibleOptions.map((opt) => (
-                        <Checkbox
-                            key={opt.id}
-                            id={opt.id}
-                            size={args.size}
-                            checked={selectedSet.has(opt.id)}
-                            onChange={(event) => handleToggle(opt.id, event.target.checked)}
-                        >
-                            {opt.label}
-                        </Checkbox>
-                    ))}
-                </CheckboxYGroup>
-            );
-
-            const renderDropdownContent = () => {
-                const renderCheckboxes = !filter.trim() || visibleOptions.length > 0;
-                return renderCheckboxes ? (
-                    renderCheckboxList()
-                ) : (
-                    <div className={`not-found ${args.size}`}>
-                        <Text size={ETextSize.B3}>Nothing was found</Text>
-                    </div>
-                );
-            };
-
-            const unselectAll = () => {
-                setSelectedIds([]);
-                setFilter("");
-            };
-
-            const handleClickClearFilter = () => {
-                unselectAll();
-            };
-
-            const handleClearInput = () => {
-                setFilter("");
-            };
-
-            const handleFilterChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-                setFilter(event.target.value);
-            };
-
-            const renderDropdown = ({ opened, setOpened, targetRef, dropdownRef }) => (
-                <MultiselectField.Dropdown
-                    opened={opened}
-                    setOpened={setOpened}
-                    targetRef={targetRef}
-                    ref={dropdownRef}
-                    focusTrapProps={{
-                        focusTrapOptions: { initialFocus: false },
-                    }}
-                    mobileViewProps={{
-                        children: (
-                            <>
-                                <DropdownMobileHeader
-                                    controlButtons={<DropdownMobileClose onClick={() => setOpened(false)} />}
-                                >
-                                    <DropdownMobileInput
-                                        placeholder="Type to proceed"
-                                        value={filter}
-                                        onChange={handleFilterChange}
-                                    />
-                                </DropdownMobileHeader>
-                                <DropdownMobileBody>{renderDropdownContent()}</DropdownMobileBody>
-                                <DropdownMobileFooter>
-                                    <Button
-                                        theme={EButtonTheme.GENERAL}
-                                        size={EComponentSize.SM}
-                                        onClick={() => setOpened(false)}
-                                    >
-                                        Button text
-                                    </Button>
-                                    <Button
-                                        theme={EButtonTheme.LINK}
-                                        size={EComponentSize.SM}
-                                        onClick={handleClickClearFilter}
-                                    >
-                                        Button link text
-                                    </Button>
-                                </DropdownMobileFooter>
-                            </>
-                        ),
-                    }}
-                >
-                    <MultiselectField.Dropdown.Header>
-                        <FormField size={args.size === EComponentSize.LG ? EComponentSize.MD : EComponentSize.SM}>
-                            <FormFieldLabel>Type to proceed</FormFieldLabel>
-                            <FormFieldInput value={filter} onChange={handleFilterChange} />
-                            <FormFieldPostfix>
-                                <FormFieldClear onClick={handleClearInput} />
-                            </FormFieldPostfix>
-                        </FormField>
-                    </MultiselectField.Dropdown.Header>
-                    <MultiselectField.Dropdown.Content loading={args.loading}>
-                        {renderDropdownContent()}
-                    </MultiselectField.Dropdown.Content>
-                    <MultiselectField.Dropdown.Footer>
-                        <Button
-                            theme={EButtonTheme.SECONDARY}
-                            size={args.size === EComponentSize.LG ? EComponentSize.MD : EComponentSize.SM}
-                            onClick={() => setOpened(false)}
-                        >
-                            Button text
-                        </Button>
-                        <Button
-                            theme={EButtonTheme.LINK}
-                            size={args.size === EComponentSize.LG ? EComponentSize.MD : EComponentSize.SM}
-                            onClick={handleClickClearFilter}
-                        >
-                            Button link text
-                        </Button>
-                    </MultiselectField.Dropdown.Footer>
-                </MultiselectField.Dropdown>
-            );
-
-            return {
-                unselectAll,
-                renderDropdown,
-                selectedIds,
-            };
-        }
-
-        const { renderDropdown, unselectAll, selectedIds } = createMultiselectFieldStoriesLogic(args);
-
-        return (
-            <ChipMultiselect
-                {...args}
-                clearSelected={unselectAll}
-                selected={selectedIds.length > 0}
-                label={args.label}
-                displayedValue={args.displayedValue}
-            >
-                {(dropdownProps) => renderDropdown(dropdownProps)}
-            </ChipMultiselect>
-        );
-    },
+    render: PlaygroundExample,
 };
 
-export const Default: StoryObj<typeof ChipMultiselect> = {
+export const Default: Story = {
     render: DefaultExample,
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: DefaultExampleSource, language: "tsx" } },
+        docs: {
+            source: {
+                code: DefaultExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
-export const Sizes: StoryObj<typeof ChipMultiselect> = {
+export const Sizes: Story = {
     render: SizesExample,
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: SizesExampleSource, language: "tsx" } },
+        docs: {
+            source: {
+                code: SizesExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
-export const Loading: StoryObj<typeof ChipMultiselect> = {
+export const Loading: Story = {
     render: LoadingExample,
     parameters: {
         controls: { disable: true },
-        docs: { source: { code: LoadingExampleSource, language: "tsx" } },
+        docs: {
+            source: {
+                code: LoadingExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
-export const VisualTests: StoryObj<typeof ChipMultiselect> = {
+export const VisualTests: Story = {
     tags: ["!autodocs"],
     parameters: {
         controls: { disable: true },
@@ -428,13 +266,7 @@ export const VisualTests: StoryObj<typeof ChipMultiselect> = {
                 }
             }, [shouldBeOpened, chipRef]);
 
-            const renderDropdown = ({
-                targetRef,
-                dropdownRef,
-            }: {
-                targetRef: React.RefObject<HTMLElement>;
-                dropdownRef: React.RefObject<HTMLDivElement>;
-            }) => (
+            const renderDropdown = ({ targetRef, dropdownRef }: ISelectExtendedFieldDropdownProvideProps) => (
                 <MultiselectField.Dropdown
                     opened={shouldBeOpened}
                     setOpened={() => {}}
