+++ stories/Chips/ChipSort.stories.tsx
@@ -1,10 +1,17 @@
 import React, { useEffect, useRef, useState } from "react";
-import { StoryObj } from "@storybook/react";
-import { ChipSort } from "../../src/components/Chip/ChipSort";
-import { EComponentSize } from "../../src/enums/EComponentSize";
+import { Meta, StoryObj } from "@storybook/react";
 import { Title, Description, Primary, Controls, Stories, Heading, ArgTypes } from "@storybook/addon-docs/blocks";
+import { ChipSort, EComponentSize, type ISelectFieldOption } from "@sberbusiness/triplex-next";
+import {
+    PlaygroundExample,
+    ChipSortPlaygroundArgs,
+    DefaultExample,
+    DefaultExampleSource,
+    SizesExample,
+    SizesExampleSource,
+} from "./examples/ChipSort";
 
-export default {
+const meta = {
     title: "Components/Chips/ChipSort",
     component: ChipSort,
     tags: ["autodocs"],
@@ -29,9 +36,13 @@ ChipSelect с иконкой выбора сортировки.
             ),
         },
     },
-};
+} satisfies Meta<typeof ChipSort>;
 
-export const Playground: StoryObj<typeof ChipSort> = {
+export default meta;
+
+type Story = StoryObj<typeof ChipSort>;
+
+export const Playground: StoryObj<ChipSortPlaygroundArgs> = {
     tags: ["!autodocs"],
     args: {
         size: EComponentSize.MD,
@@ -48,9 +59,13 @@ export const Playground: StoryObj<typeof ChipSort> = {
             control: { type: "select" },
             options: [0, 1, 2],
             description: "Дефолтное значение",
+            table: {
+                category: "Settings",
+            },
         },
     },
     parameters: {
+        testRunner: { skip: true },
         controls: {
             include: ["size", "disabled", "defaultValue"],
         },
@@ -60,85 +75,38 @@ export const Playground: StoryObj<typeof ChipSort> = {
             },
             codePanel: false,
         },
-        testRunner: { skip: true },
     },
-    render: (args) => {
-        const options = [
-            { id: "chip-sort-1", label: "По дате", value: "i1" },
-            { id: "chip-sort-2", label: "По времени", value: "i2" },
-            { id: "chip-sort-3", label: "По названию", value: "i3" },
-        ];
-
-        const [value, setValue] = useState(options[0]);
-
-        const defaultValue =
-            args.defaultValue !== undefined && args.defaultValue !== null ? options[args.defaultValue] : undefined;
-
-        return <ChipSort {...args} defaultValue={defaultValue} value={value} options={options} onChange={setValue} />;
-    },
+    render: PlaygroundExample,
 };
 
-export const Default: StoryObj<typeof ChipSort> = {
+export const Default: Story = {
+    render: DefaultExample,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        const options = [
-            { id: "chip-sort-1", label: "По дате", value: "i1" },
-            { id: "chip-sort-2", label: "По времени", value: "i2" },
-            { id: "chip-sort-3", label: "По названию", value: "i3" },
-        ];
-
-        const [value, setValue] = useState(options[0]);
-
-        return <ChipSort defaultValue={options[0]} value={value} options={options} onChange={setValue} />;
+        docs: {
+            source: {
+                code: DefaultExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
-export const Sizes: StoryObj<typeof ChipSort> = {
+export const Sizes: Story = {
+    render: SizesExample,
     parameters: {
         controls: { disable: true },
-    },
-    render: () => {
-        const options = [
-            { id: "chip-sort-1", label: "По дате", value: "i1" },
-            { id: "chip-sort-2", label: "По времени", value: "i2" },
-            { id: "chip-sort-3", label: "По названию", value: "i3" },
-        ];
-
-        const [valueSM, setValueSM] = useState(options[0]);
-        const [valueMD, setValueMD] = useState(options[0]);
-        const [valueLG, setValueLG] = useState(options[0]);
-
-        return (
-            <div style={{ display: "flex", flexDirection: "column", alignItems: "flex-start", gap: "20px" }}>
-                <ChipSort
-                    defaultValue={options[0]}
-                    size={EComponentSize.SM}
-                    value={valueSM}
-                    options={options}
-                    onChange={setValueSM}
-                />
-                <ChipSort
-                    defaultValue={options[0]}
-                    size={EComponentSize.MD}
-                    value={valueMD}
-                    options={options}
-                    onChange={setValueMD}
-                />
-                <ChipSort
-                    defaultValue={options[0]}
-                    size={EComponentSize.LG}
-                    value={valueLG}
-                    options={options}
-                    onChange={setValueLG}
-                />
-            </div>
-        );
+        docs: {
+            source: {
+                code: SizesExampleSource,
+                language: "tsx",
+            },
+        },
     },
 };
 
-export const VisualTests: StoryObj<typeof ChipSort> = {
+export const VisualTests: Story = {
+    tags: ["!autodocs"],
     parameters: {
         controls: { disable: true },
         docs: {
@@ -149,15 +117,15 @@ export const VisualTests: StoryObj<typeof ChipSort> = {
         },
     },
     render: () => {
-        const options = [
+        const options: ISelectFieldOption[] = [
             { id: "chip-sort-1", label: "По дате", value: "i1" },
             { id: "chip-sort-2", label: "По времени", value: "i2" },
             { id: "chip-sort-3", label: "По названию", value: "i3" },
         ];
 
-        const [valueSM, setValueSM] = useState(options[0]);
-        const [valueMD, setValueMD] = useState(options[0]);
-        const [valueLG, setValueLG] = useState(options[0]);
+        const [valueSM, setValueSM] = useState<ISelectFieldOption>(options[0]);
+        const [valueMD, setValueMD] = useState<ISelectFieldOption>(options[0]);
+        const [valueLG, setValueLG] = useState<ISelectFieldOption>(options[0]);
 
         const rootRef = useRef<HTMLDivElement>(null);
