+++ stories/Chips/examples/ChipMultiselect/storyConstants.ts
@@ -1,0 +1,19 @@
+export interface IStoryOption {
+    id: string;
+    label: string;
+}
+
+export const CHIP_MULTISELECT_OPTIONS: IStoryOption[] = [
+    { id: "multiselect-option-1-1", label: "Значение 1-1" },
+    { id: "multiselect-option-1-2", label: "Значение 1-2" },
+    { id: "multiselect-option-1-3", label: "Значение 1-3" },
+    { id: "multiselect-option-2-1", label: "Значение 2-1" },
+    { id: "multiselect-option-2-2", label: "Значение 2-2" },
+    { id: "multiselect-option-3", label: "Значение 3" },
+];
+
+export const CHIP_MULTISELECT_VISUAL_TEST_OPTIONS: IStoryOption[] = [
+    { id: "multiselect-option-1-1", label: "Значение 1-1" },
+    { id: "multiselect-option-2-1", label: "Значение 2-1" },
+    { id: "multiselect-option-3", label: "Значение 3" },
+];
