+++ stories/Chips/examples/ChipOptions/DefaultExample.tsx
@@ -1,0 +1,12 @@
+import React, { useState } from "react";
+import { ChipOptions } from "@sberbusiness/triplex-next";
+
+export const DefaultExample = () => {
+    const [count, setCount] = useState(0);
+
+    return (
+        <ChipOptions selected={count > 0} onClick={() => setCount((c) => c + 1)} clearSelected={() => setCount(0)}>
+            {count > 0 ? count : undefined}
+        </ChipOptions>
+    );
+};
