+++ stories/Chips/examples/ChipOptions/index.ts
@@ -2,7 +2,5 @@ export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { SizesExample } from "./SizesExample";
 export { default as SizesExampleSource } from "./SizesExample?raw";
-export { LoadingExample } from "./LoadingExample";
-export { default as LoadingExampleSource } from "./LoadingExample?raw";
 export { PlaygroundExample } from "./PlaygroundExample";
 export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
