+++ stories/Chips/examples/ChipSuggest/index.ts
@@ -2,7 +2,9 @@ export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { SizesExample } from "./SizesExample";
 export { default as SizesExampleSource } from "./SizesExample?raw";
-export { LoadingExample } from "./LoadingExample";
-export { default as LoadingExampleSource } from "./LoadingExample?raw";
+export { StatesExample } from "./StatesExample";
+export { default as StatesExampleSource } from "./StatesExample?raw";
+export { WithNotificationIconExample } from "./WithNotificationIconExample";
+export { default as WithNotificationIconExampleSource } from "./WithNotificationIconExample?raw";
 export { PlaygroundExample } from "./PlaygroundExample";
 export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
